pyproject.toml
[project]
name = "jtl-syllabus"
version = "1.20250530.2"
description = "Course management system"
readme = "README.md"
authors = [
    { name = "Eric Busboom", email = "eric.busboom@jointheleague.org" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "pydantic>=2.10.6",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/syllabus"]  # This keeps the import name as 'syllabus'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.0",
]

[project.scripts]
syl = "syllabus.cli.main:run"
